+++ src/main/java/com/ssafy/ssafsound/domain/lunch/service/LunchService.java
@@ -13,6 +13,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.Clock;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -27,10 +28,14 @@ public class LunchService {
 
     private final MetaDataConsumer metaDataConsumer;
 
+    private final Clock clock;
+
     @Transactional(readOnly = true)
     public GetLunchListResDto findLunches(Long memberId, GetLunchListReqDto getLunchListReqDto) {
 
-        Integer dayDifference = getLunchListReqDto.getDate().compareTo(LocalDate.now());
+        LocalDate currentTime = LocalDate.now(clock);
+
+        Integer dayDifference = getLunchListReqDto.getDate().compareTo(currentTime);
         if (dayDifference < 0 || dayDifference > 1) throw new LunchException(LunchErrorInfo.INVALID_DATE);
 
         MetaData campus = metaDataConsumer.getMetaData("CAMPUS",getLunchListReqDto.getCampus());
@@ -54,7 +59,7 @@ public class LunchService {
         }
 
         // 투표한 점심 메뉴 찾기
-        LunchPoll lunchPoll = lunchPollRepository.findByMember_IdAndPolledAt(memberId, LocalDate.now());
+        LunchPoll lunchPoll = lunchPollRepository.findByMember_IdAndPolledAt(memberId, currentTime);
         Long polledAt = lunchPoll != null ? (long) lunches.indexOf(lunchPoll.getLunch()) : -1L;
 
         // 인증 유저 응답
@@ -68,56 +73,6 @@ public class LunchService {
 
         return GetLunchResDto.of(lunchRepository.findById(lunchId)
                 .orElseThrow(()->new LunchException(LunchErrorInfo.INVALID_LUNCH_ID)));
-    }
-
-    @Transactional
-    public PostLunchPollResDto saveLunchPoll(Long memberId, Long lunchId) {
-
-        // 1. 점심 메뉴 존재 여부 확인
-        Lunch lunch = lunchRepository.findById(lunchId)
-                .orElseThrow(()-> new LunchException(LunchErrorInfo.INVALID_LUNCH_ID));
-
-        // 2. 점심 메뉴가 당일 메뉴인지 확인
-        if (!lunch.getCreatedAt().isEqual(LocalDate.now())) {
-            throw new LunchException(LunchErrorInfo.INVALID_DATE);
-        }
-
-        // 3. 오늘 투표한 점심 투표 엔티티 조회
-        LunchPoll lunchPoll = lunchPollRepository.findByMember_IdAndPolledAt(memberId, LocalDate.now());
-
-        // 4-1. 이미 오늘 투표한 경우
-        if (lunchPoll != null) {
-            // 5-1. 중복 투표인 경우
-            if (lunchPoll.getLunch().equals(lunch)) {
-                throw new LunchException(LunchErrorInfo.DUPLICATE_LUNCH_POLL);
-            }
-
-            // 5-2. 투표 선택지가 바뀐 경우
-            lunchPoll.setLunch(lunch);
-        }
-
-        // 4-2. 오늘 첫 투표인 경우
-        else lunchPollRepository.saveByMember_IdAndLunch_Id(memberId, lunchId, LocalDate.now());
-
-        return PostLunchPollResDto.of((long) lunch.getLunchPolls().size());
-    }
-
-    @Transactional
-    public PostLunchPollResDto deleteLunchPoll(Long memberId, Long lunchId) {
-
-        // Lunch 엔티티 조회 후 validate
-        Lunch lunch = lunchRepository.findById(lunchId)
-                .orElseThrow(() -> new LunchException((LunchErrorInfo.INVALID_LUNCH_ID)));
-
-        // 멤버 id와 Lunch 엔티티 기반으로 LunchPoll 엔티티 조회 후 validate
-        LunchPoll lunchPoll = lunchPollRepository.findByMember_IdAndLunch(memberId, lunch)
-                .orElseThrow(() -> new LunchException(LunchErrorInfo.NO_LUNCH_POLL));
-
-        // 투표 삭제
-        lunchPollRepository.delete(lunchPoll);
-
-        // 변화한 투표 수 리턴
-        return PostLunchPollResDto.of((long) lunch.getLunchPolls().size());
     }
 
 }
